+++ CHANGELOG.md
@@ -20,6 +20,10 @@ end 6.5
 
 
 
+6.4.197
+- Core: Bugfix in cronjobs new FW
+- Multi instance: All modules were always checked
+
 07-12-2020 6.4.196
 - Startpage: bugfix when removing a portlet for the second time
 - Email: Fixed download / save bug with multiple inline attachments